requirements.txt
Flask==3.0.0
easyocr==1.7.1
Levenshtein==0.27.3
Pillow==10.2.0
